Additional style guide edits
IOG style guide discourages the use of the phrase 'test ada'. The suggestion is to replace such mentions (and I found just one example) with 'testnet token'. 
This PR also suggests replacing tAda with a testnet token. If we're talking about the ticker name, should it be 'tADA' then?
This PR also suggests minor additional tweaks.